rename settings hook filename
